requirements.txt
grpcio
grpcio-tools
pygame
pytorch 
torchvision